ked/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic
pydantic-settings
deepgram-sdk
beautifulsoup4
selenium
playwright
pgvector
google-generativeai
pytest
requests
httpx
aiohttp
boto3
celery[redis]
redis
flower
numpy
